Implemented pageable readpayment funtionality and finetuned payment creation

+++ src/main/java/com/practice/payment/request/UtilityPaymentsPageRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class UtilityPaymentsPageRequest {
-    private Long pageNo;
+    private Integer pageNo;
     private Integer pageSize;
 }
